add server code sharing code with the client
